+++ packages/eslint-plugin-internal/README.md
@@ -1,5 +1,5 @@
-# eslint-plugin-internal
+# `eslint-plugin-internal`
 
-This is just a collection of internal lint rules to help enforce some guidelines specific to this repositiory.
+This is just a collection of internal lint rules to help enforce some guidelines specific to this repository.
 
 These are not intended to be used externally.
